delete

+++ exemples/delete.php
@@ -5,6 +5,11 @@ require __DIR__.'/../vendor/autoload.php';
 use Source\Models\User;
 use Source\Models\Address;
 
-$user = (new User())->findById(3);
-echo '<pre>';print_r($user);die;
-$user->destroy();
+$user = (new User())->findById(19);
+
+if($user){
+    $user->destroy();
+} else {
+    echo '<pre>';print_r($user);die;
+}
+
